pyproject.toml
[tool.poetry]
name = "posback"
version = "0.1.0"
description = ""
authors = ["josejuniorpg <yunipgx@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
django-environ = "^0.12.0"
djangorestframework = "^3.15.2"
poetry = "^2.1.1"
django-cors-headers = "^4.7.0"
django-model-utils = "^5.0.0"
pillow = "^11.1.0"
pytest-django = "^4.10.0"
factory-boy = "^3.3.3"


[tool.poetry.group.dev.dependencies]
isort = "^6.0.0"
django-extensions = "^3.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
